app/requirements.txt
langchain_openai
langchain_core
python_dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf 
chromadb
faiss-cpu
langchain_groq
unstructured==0.9.2
tiktoken==0.4.0
libmagic==1.0
python-magic==0.4.27
python-magic-bin
OpenAI == 0.28.0